+++ app/src/main/java/com/mutwakilandroiddev/go4lunch/utils/RestaurantRate.java
@@ -6,7 +6,7 @@ import android.widget.ImageView;
 public class RestaurantRate {
     // No star under 2.5, 1 star between 2.6 and 3.5, 2 stars between 3.6 and 4.5, and 3 stars above
 
-    public RestaurantRate(double rate, ImageView star1, ImageView star2, ImageView star3) {
+    public RestaurantRate(float rate, ImageView star1, ImageView star2, ImageView star3) {
         int rate_int = (int) Math.round(rate);
 
         switch (rate_int) {
